pyproject.toml
[tool.poetry]
name = "innohassle-rooms"
version = "0.1.0"
description = "A service for task management in rooms"
authors = ["Maxim Fomin <m.fomin@innopolis.university>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "~3.11"
fastapi = "^0.104.0"
uvicorn = "0.24.0.post1"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.22"}
pydantic = "^2.4.2"
pydantic-settings = "^2.0.3"
alembic = "^1.12.1"
wheel = "^0.41.3"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pytest = "^7.4.3"
httpx = "^0.25.2"
psycopg = {extras = ["binary", "pool"], version = "^3.1.15"}
sqlmodel = "^0.0.14"

[tool.poetry.group.dev.dependencies]
ruff = "^0.0.277"
black = "^23.3.0"
pre-commit = "^3.3.3"

[tool.pytest.ini_options]
pythonpath = "."
testpaths = "tests"
addopts = [
    "--strict-config",
    "--strict-markers",
]

[tool.black]
line-length = 120

[tool.ruff]
line-length = 120

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.report]
exclude_also = [
    # Don't complain about missing debug-only code:
    "def __repr__",
    "if enabled:",
    # Don't complain if tests don't hit defensive assertion code:
    "raise AssertionError",
    "raise NotImplementedError",
    # Don't complain if non-runnable code isn't run:
    "if 0:",
    "if __name__ == .__main__.:",
    # Don't complain about abstract methods, they aren't run:
    "@(abc\\.)?abstractmethod",
    # Don't complain about TYPE_CHECKING imports:
    "if TYPE_CHECKING:",
]

ignore_errors = true
